pyproject.toml
[project]
name = "gimer"
version = "1.0.0"
description = "Merge git branches with ease"
authors = [
    {name = "Shouki SAKAMOTO"}
]
dependencies = [
    "click>=8.2.1",
    "gitpython>=3.1.44",
    "inquirerpy>=0.3.4",
    "PyGithub>=2.3.0",
    "rich>=14.0.0",
]
requires-python = ">=3.10"
readme = "README.md"

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pyinstaller>=6.3.0",
    "hatchling>=1.21.0",
]

[project.scripts]
gimer = "gimer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "PL",  # pylint
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["gimer"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[dependency-groups]
dev = [
    "pyinstaller>=6.14.0",
]

[tool.poetry.dependencies]
PyGithub = "^2.3.0"
